Consolidate point.h and common.h into ristretto255.h
Folds all of the headers into a single file. Eliminates:

- Duplicated prototypes between `ristretto255.h` and `common.h`
- Duplicated field element definitions
